+++ src/shared.cpp
@@ -62,18 +62,29 @@ double lfactorial(double x) {
   return R::lgammafn(x + 1.0);
 }
 
-// Random generation for Bernoulli
+// Random generation
 
-double rng_bernoulli(double p = 0.5) {
+double rng_unif() {
+  double u;
+  // same as in base R
+  do {
+    u = R::unif_rand();
+  } while (u <= 0.0 || u >= 1.0);
+  return u;
+}
+
+double rng_bern(double p) {
   if (p < 0.0 || p > 1.0) {
     Rcpp::warning("NaNs produced");
     return NAN;
   }
-  double u = R::runif(0.0, 1.0);
-  if (u > p)
-    return 0.0;
-  else
-    return 1.0;
+  double u = rng_unif();
+  return (u > p) ? 0.0 : 1.0;
+}
+
+double rng_sign() {
+  double u = rng_unif();
+  return (u > 0.5) ? 1.0 : -1.0;
 }